requirements.txt
streamlit
langchain
pyaudio
openai
googletrans
speechrecognition
gtts
pydub
huggingface_hub
python-dotenv
Flask
